requirements.txt
pdfplumber
python-docx
Pillow
spacy
streamlit
# UI アプリ（Webベース）
streamlit